app/requirements.txt
pandas==2.0.2
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.6.0
pyspark==3.5.1
pytest==7.4.0
